Collapse api and endorser
